requirements.txt
Django>=4.1,<5.0
djangorestframework
yahoo_fin
